+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix Jira metadata collector crash on duplicate field names (OSIDB-4920)
 
+### Changed
+- Performance improvements for pghistory related queries (OSIDB-4906)
+
 ## [5.11.0] - 2026-05-28
 ### Added
 - Add validation to require PURLs (OSIDB-4810)